+++ terraform/core/catalog.tf
@@ -1,47 +1,46 @@
-resource "helm_release" "catalog-batch" {
-  repository = var.helm_repository
-  name       = "catalog-batch"
-  chart      = "${var.helm_repository}/03_catalog/01_batch"
-  version    = "1.1.1"
-  namespace  = "core"
-  create_namespace = true
-  timeout = var.helm_timeout
-
-  set {
-    name  = "image.arch"
-    value = "-native${var.server_arch}"
-  }
-  set {
-    name  = "database.password"
-    value = random_password.database_password.result
-  }
-
-}
-
 resource "helm_release" "catalog-application" {
   repository = var.helm_repository
   name       = "catalog-application"
-  chart      = "${var.helm_repository}/03_catalog/02_application"
-  version    = "1.1.1"
+  chart      = "${var.helm_repository}/catalog/application"
   namespace  = "core"
   create_namespace = true
   timeout = var.helm_timeout
 
+  set {
+    name  = "replicaCount"
+    value = local.replica_count
+  }
+  
   set {
     name  = "ingress.hosts"
     value = var.hostname
   }
   set {
     name  = "image.arch"
-    value = "-native${var.server_arch}"
+    value = "-native${local.server_arch}"
   }
   set {
     name  = "security.authentication.enabled"
     value = "false"
   }
   set {
-    name  = "replicaCount"
-    value = "1"
+    name  = "database.password"
+    value = random_password.database_password.result
+  }
+}
+
+resource "helm_release" "catalog-batch" {
+  repository = var.helm_repository
+  name       = "catalog-batch"
+  chart      = "${var.helm_repository}/catalog/batch"
+  version    = "1.1.2"
+  namespace  = "core"
+  create_namespace = true
+  timeout = var.helm_timeout
+
+  set {
+    name  = "image.arch"
+    value = "-native${local.server_arch}"
   }
   set {
     name  = "database.password"