+++ terraform/core/s3-minio.tf
@@ -1,0 +1,37 @@
+resource "helm_release" "s3-minio" {
+  name       = "s3-minio"
+  repository = "https://charts.bitnami.com/bitnami"
+  chart      = "minio"
+  namespace  = "core"
+  version    = "14.1.2"
+
+  set {
+    name  = "persistence.size"
+    value = "2Gi"
+  }
+
+  set {
+    name  = "auth.rootUser"
+    value = "minioadmin"
+  }
+  set {
+    name  = "auth.rootPassword"
+    value = "${random_password.s3_password.result}"
+  }
+  set {
+    name  = "readinessProbe.initialDelaySeconds"
+    value = "2"
+  }
+  set {
+    name  = "readinessProbe.initialDelaySeconds"
+    value = "2"
+  }
+  set {
+    name  = "extraEnvVars[0].name"
+    value = "TZ"
+  }
+  set {
+    name  = "extraEnvVars[0].value"
+    value = "Europe/Berlin"
+  }
+}