+++ terraform/core/variables.tf
@@ -1,20 +1,26 @@
 variable "hostname" {
-  default = "kubernetes"
+  default = "kind.local"
 }
 
 variable "helm_timeout" {
-  default = 60
-}
-
-variable "server_arch" {
-  default = "" #should be set from environment, only required for native images on apple silicon m1
+  default = 90
 }
 
 variable "helm_repository" {
-  default = "../../helm/templates/core" #"https://goafabric.github.io/example-konstruction-kit/helm/charts/example/spring"
+  default = "../../helm/core" #"https://goafabric.github.io/example-konstruction-kit/helm/charts/example/spring"
 }
 
-variable "authentication_enabled" {
-  default = "false"
-
+data "external" "server_arch_data" {
+  program = ["sh", "-c", "if kubectl version --output=json | grep -q 'linux/arm64'; then echo '{\"server_arch\":\"-arm64v8\"}'; else echo '{\"server_arch\":\"\"}'; fi"]
 }
+
+locals {
+  production_mode = !strcontains(var.hostname, ".local")
+
+  server_arch = data.external.server_arch_data.result["server_arch"]
+  authentication_enabled = local.production_mode
+
+  replica_count = local.production_mode ? "2" : "1"
+  postgres_ha = "false"
+}
+