+++ terraform/example/callee-service.tf
@@ -2,10 +2,14 @@ resource "helm_release" "callee-service-application" {
   repository = var.helm_repository
   name       = "callee-service-application"
   chart      = "${var.helm_repository}/callee-service/application"
-  version    = "1.1.1"
   namespace  = "example"
   create_namespace = true
   timeout = var.helm_timeout
+
+  set {
+    name  = "replicaCount"
+    value = local.replica_count
+  }
 
   set {
     name  = "ingress.hosts"
@@ -13,10 +17,6 @@ resource "helm_release" "callee-service-application" {
   }
   set {
     name  = "image.arch"
-    value = "-native${var.server_arch}"
-  }
-  set {
-    name  = "replicaCount"
-    value = "1"
+    value = "-native${local.server_arch}"
   }
 }