+++ terraform/example/person-service.tf
@@ -1,26 +1,15 @@
-resource "helm_release" "person-service-postgres" {
+resource "helm_release" "person-service-application" {
   repository = var.helm_repository
-  name       = "person-service-postgres"
-  chart      = "${var.helm_repository}/person-service/postgres"
-  version    = "1.1.1"
+  name       = "person-service-application"
+  chart      = "${var.helm_repository}/person-service/application"
   namespace  = "example"
   create_namespace = true
   timeout = var.helm_timeout
 
   set {
-    name  = "database.password"
-    value = random_password.database_password.result
+    name  = "replicaCount"
+    value = "1"
   }
-}
-
-resource "helm_release" "person-service-application" {
-  repository = var.helm_repository
-  name       = "person-service-application"
-  chart      = "${var.helm_repository}/person-service/application"
-  version    = "1.1.1"
-  namespace  = "example"
-  create_namespace = true
-  timeout = var.helm_timeout
 
   set {
     name  = "ingress.hosts"
@@ -28,10 +17,12 @@ resource "helm_release" "person-service-application" {
   }
   set {
     name  = "image.arch"
-    value = "-native${var.server_arch}"
+    value = "-native${local.server_arch}"
   }
   set {
     name  = "database.password"
     value = random_password.database_password.result
   }
 }
+
+