+++ terraform/infra/dashboard/welcome-board.tf
@@ -1,0 +1,13 @@
+resource "helm_release" "welcome-board" {
+  repository = "../../../helm/infra/welcome"
+  name       = "welcome-board"
+  chart      = "../../../helm/infra/welcome"
+  version    = "1.1.2"
+  namespace  = "default"
+  create_namespace = true
+
+  set {
+    name  = "ingress.hosts"
+    value = var.hostname
+  }
+}