+++ terraform/infra/ingress_certs/variables.tf
@@ -1,7 +1,0 @@
-variable "hostname" {
-  default = "kubernetes"
-}
-
-variable "ingress_service_type" {
-  default = "LoadBalancer"
-}