+++ terraform/infra/istio/label.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+kubectl create namespace core; kubectl label namespace core "istio.io/dataplane-mode"="ambient"
+kubectl create namespace example; kubectl label namespace example "istio.io/dataplane-mode"="ambient"