+++ terraform/infra/istio/variables.tf
@@ -1,0 +1,7 @@
+variable "profile" {
+  default = "ambient" #minimal
+}
+
+variable "hostname" {
+  default = "kind.local"
+}