+++ terraform/mesh/istio/variables.tf
@@ -1,7 +1,0 @@
-variable "profile" {
-  default = "minimal" #ambient
-}
-
-variable "hostname" {
-  default = "kubernetes"
-}