+++ xother/argocd/example/callee-service.tf
@@ -1,0 +1,45 @@
+resource "kubernetes_manifest" "callee-service-application" {
+  manifest = {
+    apiVersion = "argoproj.io/v1alpha1"
+    kind       = "Application"
+    metadata = {
+      name      = "callee-service-application"
+      namespace = "argocd"
+      finalizers = [
+        "resources-finalizer.argocd.argoproj.io",
+      ]
+    }
+    spec = {
+      project = "default"
+      source = {
+        repoURL        = var.helm_repository
+        targetRevision = "refactoring"
+        path           = "helm/example/spring/callee-service/application"
+        helm = {
+          parameters = [
+            {
+              name  = "ingress.hosts"
+              value = var.hostname
+            },
+            {
+              name  = "image.arch"
+              value = "-native${local.server_arch}"
+            },
+            {
+              name  = "replicaCount"
+              value = "1"
+            },
+          ]
+        }
+      }
+      destination = {
+        server    = "https://kubernetes.default.svc"
+        namespace = "example"
+      }
+      syncPolicy = {
+        automated  = {}
+        syncOptions = ["CreateNamespace=true"]
+      }
+    }
+  }
+}